split structure of unicorn parser
